Moficaciones y mejores a idea-CMWP
Modifique el codigo de idea_cmwp.c de modo tal de poder elegir
libremente el ancho y la cantidad de porciones del anillo de Debye con
el que voy a trabajar. Ahora hay que especificar el ancho de la porcion
del anillo de Debye que se quiere promediar, asi como la separacion
angular entre anillos.
Modifique la forma en que la rutina CMWP fitting presenta el progreso de
los ajustes. Ahora, el ajuste por patron se muestra siempre en el mismo
renglon y solo se pasa a una nueva linea cuando se cambia de spr.

+++ idea-CMWP/pv.c
@@ -8,7 +8,7 @@
 #include "aux_functions.h"
 #include "pv_steps.c"
 
-void pv_fitting(int exists, exp_data * sync_data, peak_data * difra, float * intens, double ** seeds)
+void pv_fitting(int exists, exp_data * sync_data, peak_data * difra, double * intens, double ** seeds)
 {
     //printf("Inicio pv_fitting\n");
     //variables auxiliares del programa
